Refactor HomeView and LoginView to use new SnackBar extension and reorganize widget structure
- Introduce a new SnackBar extension for BuildContext to simplify the process of showing snack bars across the application.
- Move the HomeViewBody and HomeTableBody widgets into their respective files under the widgets directory to improve code organization and maintainability.
- Update imports in HomeView and LoginView to reference the corrected path for the SnackBar extension.
- Remove redundant code from HomeView and delegate the responsibility of building the home view body to the newly created HomeViewBody widget.
- Create a new HomeTableBody widget to encapsulate the logic for building the table body in the home view, enhancing readability and separation of concerns.
- Ensure all references to the SnackBar extension are updated to use the correct path, maintaining consistency across the codebase.
