+++ lib/features/login/ui/view/login_view.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
 import 'package:modal_progress_hud_nsn/modal_progress_hud_nsn.dart';
-import 'package:to_allah/core/extesions/snack_bar_extension.dart';
+import 'package:to_allah/core/extensions/snack_bar_extension.dart';
 import 'package:to_allah/core/routing/app_router.dart';
 import 'package:to_allah/features/login/cubits/login_cubit.dart';
